Add skia and update freya
