+++ crates/ui/Cargo.toml
@@ -16,3 +16,4 @@ synk_core = { path = "../core", package = "core" }
 tree-sitter = "0.20.10"
 tree-sitter-rust = "0.20.4"
 cc = "*"
+skia-safe = "0.70.0"
